Moved validation results inside new sub-namespace.
