+++ UIPathValidator/Validation/Result/ArgumentValidationResult.cs
@@ -1,6 +1,6 @@
 using UIPathValidator.UIPath;
 
-namespace UIPathValidator.Validation
+namespace UIPathValidator.Validation.Result
 {
     public class ArgumentValidationResult : ValidationResult
     {
