+++ UIPathValidator/Validation/Result/UnreachableWorkflowValidationResult.cs
@@ -1,6 +1,6 @@
 using UIPathValidator.UIPath;
 
-namespace UIPathValidator.Validation
+namespace UIPathValidator.Validation.Result
 {
     public class UnreachableWorkflowValidationResult : ValidationResult
     {
