+++ UIPathValidator/Validation/Result/ValidationResultType.cs
@@ -1,4 +1,4 @@
-namespace UIPathValidator.Validation
+namespace UIPathValidator.Validation.Result
 {
     public enum ValidationResultType
     {
