+++ UIPathValidator/Validation/Validator.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using UIPathValidator.Validation.Result;
 
 namespace UIPathValidator.Validation
 {
